requirements.txt
Flask
redis
shortuuid
python-dotenv
gunicorn
